Task 12/requirements.txt
flask>=2.3.0
sentence-transformers==2.7.0
faiss-cpu>=1.7.4
numpy>=1.24.0
